+++ src/mixin/mod.rs
@@ -4,6 +4,7 @@ use crate::{
     prelude::*,
     ruby::{self, ID, VALUE},
     util::Sealed,
+    vm::EvalArgs,
 };
 
 mod class;
@@ -323,7 +324,7 @@ pub trait Mixin: Object + Sealed {
     /// An exception may be raised by the code or by `args` being invalid.
     #[inline]
     unsafe fn eval(self, args: impl EvalArgs) -> AnyObject {
-        args.eval_in(self)
+        args.eval_in_mixin(self)
     }
 
     /// Evaluates `args` in the context of `self`, returning any raised
@@ -339,7 +340,7 @@ pub trait Mixin: Object + Sealed {
     /// type `A`.
     #[inline]
     unsafe fn eval_protected(self, args: impl EvalArgs) -> Result<AnyObject> {
-        args.eval_in_protected(self)
+        args.eval_in_mixin_protected(self)
     }
 }
 
@@ -364,133 +365,6 @@ impl Mixin for Module {
     #[inline]
     fn to_module(self) -> Result<Module, Class> {
         Ok(self)
-    }
-}
-
-/// A type that can be used as one or more arguments for evaluating code within
-/// the context of a [`Mixin`](trait.Mixin.html).
-///
-/// See the documentation of [its implementors](#foreign-impls) for much more
-/// detailed information.
-pub trait EvalArgs: Sized {
-    /// Evaluates `self` in the context of `mixin`, returning any thrown
-    /// exceptions.
-    ///
-    /// # Safety
-    ///
-    /// Code executed from `self` may void the type safety of objects accessible
-    /// from Rust. For example, if one calls `push` on an `Array<A>` with an
-    /// object of type `B`, then the inserted object will be treated as being of
-    /// type `A`.
-    #[inline]
-    unsafe fn eval_in_protected(self, mixin: impl Mixin) -> Result<AnyObject>;
-
-    /// Evaluates `self` in the context of `mixin`.
-    ///
-    /// # Safety
-    ///
-    /// Code executed from `self` may void the type safety of objects accessible
-    /// from Rust. For example, if one calls `push` on an `Array<A>` with an
-    /// object of type `B`, then the inserted object will be treated as being of
-    /// type `A`.
-    ///
-    /// If an exception is thrown due to an argument error or from evaluating
-    /// the script itself, it should be caught.
-    unsafe fn eval_in(self, mixin: impl Mixin) -> AnyObject;
-}
-
-/// Unchecked arguments directly to the evaluation function.
-impl<O: Object> EvalArgs for &[O] {
-    #[inline]
-    unsafe fn eval_in_protected(self, mixin: impl Mixin) -> Result<AnyObject> {
-        // monomorphization
-        unsafe fn eval(args: &[AnyObject], mixin: VALUE) -> Result<AnyObject> {
-            let raw = crate::protected_no_panic(|| ruby::rb_mod_module_eval(
-                args.len() as _,
-                args.as_ptr() as *const ruby::VALUE,
-                mixin,
-            ))?;
-            Ok(AnyObject::from_raw(raw))
-        }
-        eval(AnyObject::convert_slice(self), mixin.raw())
-    }
-
-    #[inline]
-    unsafe fn eval_in(self, mixin: impl Mixin) -> AnyObject {
-        let raw = ruby::rb_mod_module_eval(
-            self.len() as _,
-            self.as_ptr() as *const ruby::VALUE,
-            mixin.raw(),
-        );
-        AnyObject::from_raw(raw)
-    }
-}
-
-/// The script argument without any extra information.
-impl EvalArgs for String {
-    #[inline]
-    unsafe fn eval_in_protected(self, mixin: impl Mixin) -> Result<AnyObject> {
-        self.as_any_slice().eval_in_protected(mixin)
-    }
-
-    #[inline]
-    unsafe fn eval_in(self, mixin: impl Mixin) -> AnyObject {
-        self.as_any_slice().eval_in(mixin)
-    }
-}
-
-/// The script argument as a UTF-8 string, without any extra information.
-// TODO: Impl for `Into<String>` when specialization stabilizes
-impl EvalArgs for &str {
-    #[inline]
-    unsafe fn eval_in_protected(self, mixin: impl Mixin) -> Result<AnyObject> {
-        String::from(self).eval_in_protected(mixin)
-    }
-
-    #[inline]
-    unsafe fn eval_in(self, mixin: impl Mixin) -> AnyObject {
-        String::from(self).eval_in(mixin)
-    }
-}
-
-/// The script and filename arguments.
-impl<S: Into<String>, F: Into<String>> EvalArgs for (S, F) {
-    #[inline]
-    unsafe fn eval_in_protected(self, mixin: impl Mixin) -> Result<AnyObject> {
-        let (s, f) = self;
-        [s.into(), f.into()].eval_in_protected(mixin)
-    }
-
-    #[inline]
-    unsafe fn eval_in(self, mixin: impl Mixin) -> AnyObject {
-        let (s, f) = self;
-        [s.into(), f.into()].eval_in(mixin)
-    }
-}
-
-/// The script, filename, and line number arguments.
-impl<S, F, L> EvalArgs for (S, F, L)
-where
-    S: Into<String>,
-    F: Into<String>,
-    L: Into<Integer>,
-{
-    #[inline]
-    unsafe fn eval_in_protected(self, mixin: impl Mixin) -> Result<AnyObject> {
-        let (s, f, l) = self;
-        let s = AnyObject::from(s.into());
-        let f = AnyObject::from(f.into());
-        let l = AnyObject::from(l.into());
-        [s, f, l].eval_in_protected(mixin)
-    }
-
-    #[inline]
-    unsafe fn eval_in(self, mixin: impl Mixin) -> AnyObject {
-        let (s, f, l) = self;
-        let s = AnyObject::from(s.into());
-        let f = AnyObject::from(f.into());
-        let l = AnyObject::from(l.into());
-        [s, f, l].eval_in(mixin)
     }
 }