+++ src/vm/mod.rs
@@ -2,7 +2,6 @@
 
 use std::{
     error::Error,
-    ffi::CStr,
     fmt,
     num::NonZeroI32,
     os::raw::c_int,
@@ -12,8 +11,13 @@ use crate::{
     ruby,
 };
 
+mod eval;
 mod instr_seq;
-pub use instr_seq::*;
+
+pub use self::{
+    eval::*,
+    instr_seq::*,
+};
 
 /// Initializes the Ruby VM, returning an error code if it failed.
 #[inline]
@@ -218,108 +222,6 @@ pub unsafe fn load_protected(file: impl Into<String>, wrap: bool) -> Result {
     ruby::rb_load_protect(file.into().raw(), wrap as c_int, &mut err);
     match err {
         0 => Ok(()),
-        _ => Err(AnyException::_take_current()),
-    }
-}
-
-/// Evaluates `script` in an isolated binding without handling exceptions.
-///
-/// Variables:
-/// - `__FILE__`: "(eval)"
-/// - `__LINE__`: starts at 1
-///
-/// # Safety
-///
-/// Code executed from `script` may void the type safety of objects accessible
-/// from Rust. For example, if one calls `push` on `Array<A>` with an object of
-/// type `B`, then the inserted object will be treated as being of type `A`.
-///
-/// ```
-/// # rosy::vm::init().unwrap();
-/// use std::ffi::CStr;
-/// use rosy::prelude::*;
-///
-/// let array: Array<Integer> = (1..=3).collect(); // [1, 2, 3]
-///
-/// let module = Module::def("Evil").unwrap();
-/// unsafe { module.set_const("ARR", array) };
-///
-/// let script = b"Evil::ARR.push('hi')\0";
-/// let script = CStr::from_bytes_with_nul(script).unwrap();
-///
-/// unsafe { rosy::vm::eval(script) };
-/// let hi = array.get(3).unwrap();
-/// ```
-///
-/// If we try using `hi` as an `Integer` here, we will get a segmentation fault:
-///
-// Supports all failures, apparently
-/// ```should_panic
-// This is just for demonstration purposes
-/// # rosy::vm::init().unwrap();
-/// # use rosy::prelude::*;
-/// # let hi = unsafe { Integer::cast_unchecked(String::from("hi")) };
-/// let val = hi.to_truncated::<i32>();
-/// ```
-///
-/// However, we can see that `hi` is actually a `String` despite being typed as
-/// an `Integer`:
-///
-/// ```
-// This is just for demonstration purposes
-/// # rosy::vm::init().unwrap();
-/// # use rosy::prelude::*;
-/// # let hi = AnyObject::from("hi");
-/// let hi = unsafe { String::cast_unchecked(hi) };
-/// assert_eq!(hi, "hi");
-/// ```
-///
-/// ...also, any exceptions raised in `script` must be handled in Rust-land.
-#[inline]
-pub unsafe fn eval(script: &CStr) -> AnyObject {
-    AnyObject::from_raw(ruby::rb_eval_string(script.as_ptr()))
-}
-
-/// Evaluates `script` in an isolated binding, returning an exception if one is
-/// raised.
-///
-/// Variables:
-/// - `__FILE__`: "(eval)"
-/// - `__LINE__`: starts at 1
-///
-/// # Safety
-///
-/// Code executed from `script` may void the type safety of objects accessible
-/// from Rust. For example, if one calls `push` on `Array<A>` with an object of
-/// type `B`, then the inserted object will be treated as being of type `A`.
-#[inline]
-pub unsafe fn eval_protected(script: &CStr) -> Result<AnyObject> {
-    let mut err = 0;
-    let raw = ruby::rb_eval_string_protect(script.as_ptr(), &mut err);
-    match err {
-        0 => Ok(AnyObject::from_raw(raw)),
-        _ => Err(AnyException::_take_current()),
-    }
-}
-
-/// Evaluates `script` under a module binding in an isolated binding, returning
-/// an exception if one is raised.
-///
-/// Variables:
-/// - `__FILE__`: "(eval)"
-/// - `__LINE__`: starts at 1
-///
-/// # Safety
-///
-/// Code executed from `script` may void the type safety of objects accessible
-/// from Rust. For example, if one calls `push` on `Array<A>` with an object of
-/// type `B`, then the inserted object will be treated as being of type `A`.
-#[inline]
-pub unsafe fn eval_wrapped(script: &CStr) -> Result<AnyObject> {
-    let mut err = 0;
-    let raw = ruby::rb_eval_string_wrap(script.as_ptr(), &mut err);
-    match err {
-        0 => Ok(AnyObject::from_raw(raw)),
         _ => Err(AnyException::_take_current()),
     }
 }